Move allfields model to a different app
