+++ tests/test_app/utils/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+from django.conf import settings
+
+
+class UtilsConfig(AppConfig):
+    name = f"{settings.PREFIX}utils"
